+++ src/components/TodoList/TodoList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Todo } from '../../types';
-import { UserInfo } from '../UserInfo';
+import { TodoInfo } from '../TodoInfo';
 
 type Props = {
   todos: Todo[];
@@ -9,15 +9,8 @@ type Props = {
 export const TodoList: React.FC<Props> = ({ todos }) => {
   return (
     <section className="TodoList">
-      {todos.map(todo => (
-        <article
-          key={todo.id}
-          data-id={todo.id}
-          className={`TodoInfo ${todo.completed ? 'TodoInfo--completed' : ''}`}
-        >
-          <h2 className="TodoInfo__title">{todo.title}</h2>
-          {todo.user && <UserInfo user={todo.user} />}
-        </article>
+      {todos.map((todo: Todo) => (
+        <TodoInfo todo={todo} key={todo.id} />
       ))}
     </section>
   );
